fix global fitting not setting Oscillation fit type & segment, also change all literals to enums for better runtime validation.
